feat: update_cwd cache
